src/llm_rag/search.py
from typing import Optional
import argparse
import asyncio

import lancedb
import google.generativeai as genai
from mcp import Server, Request, Response

from .types import SearchResult

class SearchServer:
    def __init__(self, db_path: str, table_name: str):
        """Initialize search server.
        
        Args:
            db_path: Path to LanceDB database
            table_name: Name of table containing embeddings
        """
        self.db = lancedb.connect(db_path)
        self.table = self.db[table_name]

    async def search(
        self,
        query: str,
        limit: int = 10,
        score_threshold: float = 0.0,
    ) -> list[SearchResult]:
        """Search for documents matching query.
        
        Args:
            query: Search query
            limit: Maximum number of results
            score_threshold: Minimum similarity score
            
        Returns:
            List of search results
        """
        # TODO: Implement hybrid search using Gemini embeddings and LanceDB
        raise NotImplementedError

async def handle_request(request: Request, server: SearchServer) -> Response:
    """Handle incoming search requests."""
    try:
        query = request.json.get("query")
        if not query:
            return Response({"error": "Missing query"}, status=400)

        limit = request.json.get("limit", 10)
        score_threshold = request.json.get("score_threshold", 0.0)

        results = await server.search(query, limit, score_threshold)
        return Response({"results": [r.dict() for r in results]})
    except Exception as e:
        return Response({"error": str(e)}, status=500)

def main():
    parser = argparse.ArgumentParser(description="Run search server")
    parser.add_argument("--db", required=True, help="Path to LanceDB database")
    parser.add_argument(
        "--table",
        default="documents",
        help="LanceDB table name",
    )
    parser.add_argument(
        "--host",
        default="localhost",
        help="Host to bind server to",
    )
    parser.add_argument(
        "--port",
        type=int,
        default=8080,
        help="Port to bind server to",
    )
    args = parser.parse_args()

    # Initialize search server
    server = SearchServer(args.db, args.table)

    # Start MCP server
    mcp_server = Server()
    mcp_server.handle("/search", lambda req: handle_request(req, server))
    asyncio.run(mcp_server.serve(args.host, args.port))

if __name__ == "__main__":
    main()
