pyproject.toml
[project]
name = "llm_rag"
version = "0.1.0"
description = "RAG implementation using LlamaIndex, Gemini, and LanceDB"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters",
    "aiohttp",
    "google-generativeai>=0.3.0",
    "lancedb>=0.4.0",
    "mcp>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "pydantic>=2.0.0",
    "structlog>=24.1.0",
    "tabulate>=0.9.0",
    "markdownify",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
